refactor: typeを移動
