+++ src/view/plancandidate/DialogRelatedPlaces.tsx
@@ -30,6 +30,7 @@ import { PlacesWithCategory } from "src/domain/models/PlacesWithCategory";
 import { Transition } from "src/domain/models/Transition";
 import { copyObject } from "src/domain/util/object";
 import { useAppTranslation } from "src/hooks/useAppTranslation";
+import { OnClickHandler } from "src/types/handler";
 import { FullscreenDialog } from "src/view/common/FullscreenDialog";
 import { ImageSliderPreview } from "src/view/common/ImageSliderPreview";
 import { RoundedButton } from "src/view/common/RoundedButton";
@@ -39,7 +40,6 @@ import {
     PlaceChipActionInstagram,
 } from "src/view/plandetail/PlaceChipContextAction";
 import { PlaceInfoTab } from "src/view/plandetail/PlaceInfoTab";
-import { OnClickHandler } from "src/view/types/handler";
 
 type Props = {
     visible: boolean;
