+++ src/view/plandetail/PlaceChipContextAction.tsx
@@ -13,7 +13,7 @@ import { SiGooglemaps, SiInstagram } from "react-icons/si";
 import { AnalyticsEvents } from "src/constant/analytics";
 import { useAppTranslation } from "src/hooks/useAppTranslation";
 import { UploadPlaceImageProps } from "src/hooks/useUploadPlaceImage";
-import { OnClickHandler } from "src/view/types/handler";
+import { OnClickHandler } from "src/types/handler";
 
 type Props = {
     label: string;
